conflict file changed in b2

+++ app/src/main/java/com/farizma/truthdare/conflict.java
@@ -1,7 +1,7 @@
 package com.farizma.truthdare;
 
 public class conflict {
-    int roll;
+    double roll;
     String id,name;
 
     public conflict(int roll, String id, String name) {
